matam/requirements.txt
async-timeout==5.0.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
dlib==20.0.0
dnspython==2.7.0
eventlet==0.40.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
gdown==5.2.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rq==2.4.0
rsa==4.9.1
simple-websocket==1.1.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
